+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-CkBQNbd2.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-Fa484Y6w.js";import{_ as a}from"./mermaid.core-VYhUct1A.js";import"./chunk-5VM5RSS4-kuWTPvCB.js";import"./chunk-XXDRQBXY-D8aQxzT0.js";import"./chunk-POPQ4Y6H-B4cLwp2x.js";import"./chunk-F27PBJKO-7Swcs57Q.js";import"./channel-Cz2kSHpg.js";import"./index-CUoOjPvQ.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};