+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-D5ml4Y80.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-4C-k8m1U.js";import{_ as a}from"./mermaid.core-D9E_Xc6f.js";import"./chunk-5VM5RSS4-D1TJQJJm.js";import"./chunk-XXDRQBXY-DtagCtoo.js";import"./chunk-POPQ4Y6H-DFyh2a7R.js";import"./chunk-F27PBJKO-D-v6_TRV.js";import"./channel-DNJcSkBx.js";import"./index-CEPS3jt3.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};